workspace/flask/requirements.txt
flask>=2.2.0
werkzeug>=2.3.0
# langchain
# langchain_openai
intersystems-irispython
iris
numpy
sentence_transformers

# llama-index>=0.10.0
# llama-index-llms-openai
pandas
hf_xet
diskcache
openai
python-dotenv

